+++ bindings/netstandard/ElectionGuard/ElectionGuard.Decryption/Tally/TallyAccumulationResult.cs
@@ -1,5 +1,5 @@
 using System.Text;
-using ElectionGuard.Encryption.Ballot;
+using ElectionGuard.Ballot;
 
 namespace ElectionGuard.Decryption.Tally;
