+++ bindings/netstandard/ElectionGuard/ElectionGuard.ElectionSetup/ElectionPolynomial.cs
@@ -1,7 +1,7 @@
-﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations;
 using ElectionGuard.ElectionSetup.Extensions;
 using ElectionGuard.Proofs;
-using ElectionGuard.UI.Lib.Extensions;
+using ElectionGuard.Extensions;
 using Newtonsoft.Json;
 
 namespace ElectionGuard.ElectionSetup;
