+++ bindings/netstandard/ElectionGuard/ElectionGuard.ElectionSetup/Guardian/Guardian.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics;
 using ElectionGuard.ElectionSetup.Extensions;
 using ElectionGuard.Proofs;
-using ElectionGuard.UI.Lib.Extensions;
+using ElectionGuard.Extensions;
 using ElectionGuard.UI.Lib.Models;
 
 namespace ElectionGuard.ElectionSetup;
